add entity, dto, repository,service, controller billing

+++ src/main/java/com/example/yootday1/dto/scheduleSlot/ScheduleSlotResponse.java
@@ -9,7 +9,7 @@ public class ScheduleSlotResponse {
 
     private String slotCode;
 
-    private Integer weekday;
+    private byte weekday;
 
     private LocalTime startTime;
